+++ flutter_feature_biometric_ios/lib/src/constant/error_constant.dart
@@ -1,0 +1,8 @@
+// ignore_for_file: constant_identifier_names
+class ErrorConstantIOS {
+  ErrorConstantIOS._();
+
+  static const IOS_UNKNOWN_POLICY = 'IOS_UNKNOWN_POLICY';
+  static const IOS_UNKNOWN_RESULT = 'IOS_UNKNOWN_RESULT';
+  static const IOS_UNKNOWN_UNABLE_AUTHENTICATE = 'IOS_UNKNOWN_UNABLE_AUTHENTICATE';
+}