+++ flutter_feature_biometric_platform_interface/lib/src/enum/biometric_authenticate_status.dart
@@ -1,6 +1,0 @@
-enum BiometricAuthenticateStatus {
-  success,
-  failed,
-  error,
-  dialogClicked,
-}