+++ flutter_feature_biometric_platform_interface/lib/src/enum/biometric_authenticator_type.dart
@@ -1,4 +1,0 @@
-enum BiometricAuthenticatorType {
-  biometric,
-  deviceCredential
-}