+++ flutter_feature_biometric_platform_interface/lib/src/exception/feature_biometric_exception.dart
@@ -1,0 +1,9 @@
+class FeatureBiometricException {
+  String code;
+  String? message;
+
+  FeatureBiometricException({
+    required this.code,
+    this.message,
+  });
+}