+++ flutter_feature_biometric_ios/CHANGELOG.md
@@ -1,6 +1,7 @@
-## 0.0.2
+## 0.1.0
 
 * Update docs
+* Update ios using platform interface 0.1.0
 
 
 ## 0.0.1